Fixed warning

+++ src/TDataTaking.cpp
@@ -136,7 +136,7 @@ void TDataTaking::FindTimeoutHics (int iboard, int *triggerCounts, int nTriggers
       }
     }
     // HIC is connected to this readout board AND did not send enough events
-    if ((isOnBoard) && (nTrigs < nTriggers * m_hics.at(iHic)->GetNEnabledChips())) {
+    if ((isOnBoard) && (nTrigs < nTriggers * (int) (m_hics.at(iHic)->GetNEnabledChips()))) {
       m_errorCounts.at(m_hics.at(iHic)->GetDbId()).nTimeout ++;
     }
   }
